pyproject.toml
[project]
name = "tokopedia-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crawlee[all]>=0.6.12",
    "fastapi>=0.104.1",
    "pydantic>=2.5.0",
    "requests>=2.32.5",
    "uvicorn[standard]>=0.35.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.11",
]
